used onWillFocus to load chat list in chat list screen
when navigating to chat, the chat list screen was updated since the chat
screen component is mounted first then the chat screen is mounted.
Now when navigating to chat screen, the updateChatList is not called.
updateChatLIst function is moved inside the onWillFoucs.
So when navigating to chat screen, the chat list screen is not focused,
so no call to updateChatList function.